pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Python task-manager on django"
authors = ["a.cheremushkin <a.ch3re359@yandex.ru>"]
readme = "README.md"
packages = []


[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
dj-database-url = "^1.0.0"
python-dotenv = "^0.21.0"
django-bootstrap4 = "^22.2"
gunicorn = "^20.1.0"
django-filter = "^22.1"

[tool.poetry.group.dev.dependencies]
loguru = "^0.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
